+++ src/output/displayManager/DisplayManager.cpp
@@ -156,6 +156,48 @@ void DisplayManager::DisplayEmblemCount()
         NJM_LOCATION(2, this->_startLine - 1), ("Emblems: " + std::to_string(_emblemCount)).c_str());
 }
 
+std::string DisplayManager::GetLevelTime(const bool showTarget)
+{
+    int targetTime = 0;
+
+    switch (CurrentCharacter)
+    {
+    case Characters_Sonic:
+        if (SONIC_TARGET_TIMES.find(CurrentLevel) != SONIC_TARGET_TIMES.end())
+            targetTime = SONIC_TARGET_TIMES.at(CurrentLevel);
+        break;
+    case Characters_Tails:
+        if (TAILS_TARGET_TIMES.find(CurrentLevel) != TAILS_TARGET_TIMES.end())
+            targetTime = TAILS_TARGET_TIMES.at(CurrentLevel);
+        break;
+    case Characters_Knuckles:
+        if (KNUCKLES_TARGET_TIMES.find(CurrentLevel) != KNUCKLES_TARGET_TIMES.end())
+            targetTime = KNUCKLES_TARGET_TIMES.at(CurrentLevel);
+        break;
+    case Characters_Amy:
+        if (AMY_TARGET_TIMES.find(CurrentLevel) != AMY_TARGET_TIMES.end())
+            targetTime = AMY_TARGET_TIMES.at(CurrentLevel);
+        break;
+    case Characters_Gamma:
+        if (GAMMA_TARGET_TIMES.find(CurrentLevel) != GAMMA_TARGET_TIMES.end())
+            targetTime = GAMMA_TARGET_TIMES.at(CurrentLevel);
+        break;
+    case Characters_Big:
+        return showTarget ? " 2000 " : "(    )";
+
+    default: return "";
+    }
+    const int minutes = targetTime / 60 / 60;
+    const int seconds = targetTime / 60 % 60;
+    const std::string formattedTime = " " + std::to_string(minutes) + ":" + (seconds < 10 ? "0" : "") +
+        std::to_string(seconds) + " ";
+
+    if (showTarget)
+        return formattedTime;
+    else
+        return "(" + std::string(formattedTime.length() - 2, ' ') + ")";
+}
+
 
 void DisplayManager::OnEnterCharacterSelectScreen()
 {
@@ -173,10 +215,11 @@ void DisplayManager::UpdateChecks(const std::map<int, LocationData>& checkData)
     this->_checkData = checkData;
 }
 
-void DisplayManager::SetMessageConfiguration(const float messageDisplayDuration, const int messageFontSize, const int messageColor)
+void DisplayManager::SetMessageConfiguration(const float messageDisplayDuration, const int messageFontSize,
+                                             const int messageColor)
 {
     this->_displayDuration = messageDisplayDuration;
-    this->_debugFontSize =  messageFontSize;
+    this->_debugFontSize = messageFontSize;
     this->_displayMessageColor = messageColor;
 }
 
@@ -255,12 +298,15 @@ void DisplayManager::DisplayItemsUnlocked()
         displayOffset++;
         buffer.clear();
         buffer.append(" ");
+        const std::string levelTime = "2:45";
         if (missionsEnabled > 0)
             buffer.append(GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_C) ? "C" : " ");
         if (missionsEnabled > 1)
             buffer.append(GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_B) ? " B" : "  ");
         if (missionsEnabled > 2)
-            buffer.append(GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_A) ? " A" : "  ");
+            buffer.append(GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_A)
+                              ? " A" + this->GetLevelTime(true)
+                              : "  " + this->GetLevelTime(true));
         buffer.append(" ");
 
         if (_options.lifeSanity && _options.GetCharacterLifeSanity(static_cast<Characters>(CurrentCharacter)))
@@ -287,7 +333,9 @@ void DisplayManager::DisplayItemsUnlocked()
         if (missionsEnabled > 1)
             buffer.append(!GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_B) ? "-B" : "- ");
         if (missionsEnabled > 2)
-            buffer.append(!GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_A) ? "-A" : "- ");
+            buffer.append(!GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_A)
+                              ? "-A" + this->GetLevelTime(false)
+                              : "- " + this->GetLevelTime(false));
         buffer.append("]");
 
         SetDebugFontColor(currentColor & 0x00FFFFFF | 0x66000000);